@@ -24,76 +24,67 @@
     "url": "https://github.com/squidfunk/mkdocs-material.git"
   },
   "scripts": {
-    "build": "npm run clean && npx webpack --mode production",
-    "clean": "npx rimraf material",
-    "lint": "npm run lint:ts && npm run lint:scss",
-    "lint:scss": "npx stylelint \"src/assets/**/*.scss\"",
-    "lint:ts": "npx tslint -p tsconfig.json \"src/**/*.ts\"",
-    "start": "npx webpack --mode development --watch"
+    "build": "rimraf material && ts-node -T tools/build --optimize",
+    "build:dirty": "ts-node -T tools/build --dirty",
+    "clean": "rimraf material",
+    "lint": "npm run lint:scss && npm run lint:ts",
+    "lint:scss": "stylelint \"src/assets/**/*.scss\"",
+    "lint:ts": "eslint --cache \"src/**/*.ts\"",
+    "start": "ts-node -T tools/build --verbose --dirty --watch"
   },
   "dependencies": {
-    "clipboard": "^2.0.6",
+    "clipboard": "^2.0.8",
     "escape-html": "^1.0.3",
-    "lunr": "^2.3.8",
+    "focus-visible": "^5.2.0",
+    "fuzzaldrin-plus": "^0.6.0",
+    "lunr": "^2.3.9",
     "lunr-languages": "^1.4.0",
-    "ramda": "^0.27.0",
-    "rxjs": "^7.0.0-beta.0"
+    "rxjs": "^7.0.0"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.13.0",
-    "@mdi/svg": "^5.3.45",
-    "@primer/octicons": "^9.6.0",
+    "@fortawesome/fontawesome-free": "^5.15.3",
+    "@mdi/svg": "^5.9.55",
+    "@primer/octicons": "^13.0.0",
     "@types/clipboard": "^2.0.1",
-    "@types/copy-webpack-plugin": "^6.0.0",
-    "@types/escape-html": "1.0.0",
-    "@types/event-hooks-webpack-plugin": "^2.0.0",
-    "@types/html-minifier": "^3.5.3",
+    "@types/escape-html": "1.0.1",
+    "@types/fuzzaldrin-plus": "^0.6.1",
+    "@types/html-minifier": "^4.0.0",
     "@types/lunr": "^2.3.3",
-    "@types/mini-css-extract-plugin": "^0.9.1",
-    "@types/node": "^14.0.11",
-    "@types/ramda": "^0.27.6",
-    "@types/resize-observer-browser": "^0.1.3",
-    "@types/webpack": "^4.41.17",
-    "@types/webpack-assets-manifest": "^3.0.1",
-    "autoprefixer": "^9.8.0",
-    "copy-webpack-plugin": "^6.0.2",
-    "css-loader": "^3.5.3",
-    "event-hooks-webpack-plugin": "^2.1.6",
-    "expose-loader": "^0.7.5",
+    "@types/node": "^15.0.1",
+    "@types/resize-observer-browser": "^0.1.5",
+    "@types/sass": "^1.16.0",
+    "@typescript-eslint/eslint-plugin": "^4.22.0",
+    "@typescript-eslint/parser": "^4.22.0",
+    "autoprefixer": "^10.2.5",
+    "chokidar": "^3.5.1",
+    "cssnano": "^5.0.2",
+    "esbuild": "^0.11.18",
+    "eslint": "^7.25.0",
+    "eslint-plugin-eslint-comments": "^3.2.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsdoc": "^33.0.0",
+    "eslint-plugin-no-null": "^1.0.2",
     "github-types": "^1.0.0",
     "gitlab": "^14.2.2",
     "html-minifier": "^4.0.0",
-    "ignore-emit-webpack-plugin": "^2.0.2",
-    "imagemin-webpack-plugin": "^2.4.2",
     "material-design-color": "^2.3.2",
     "material-shadows": "^3.0.1",
-    "mini-css-extract-plugin": "^0.9.0",
-    "modularscale-sass": "^3.0.10",
-    "postcss-inline-svg": "^4.1.0",
-    "postcss-loader": "^3.0.0",
-    "postcss-svgo": "^4.0.2",
-    "preact": "^10.4.4",
-    "resize-observer-polyfill": "^1.5.1",
+    "postcss": "^8.2.13",
+    "postcss-inline-svg": "^5.0.0",
+    "preact": "^10.5.13",
     "rimraf": "^3.0.2",
-    "sass": "^1.26.8",
-    "sass-loader": "^8.0.2",
-    "stylelint": "^13.6.0",
+    "sass": "^1.32.12",
+    "stylelint": "^13.13.1",
     "stylelint-config-rational-order": "^0.1.2",
-    "stylelint-config-standard": "^20.0.0",
-    "stylelint-order": "^4.1.0",
-    "stylelint-scss": "^3.17.2",
-    "ts-loader": "^7.0.5",
-    "ts-node": "^8.10.2",
-    "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "tslib": "^2.0.0",
-    "tslint": "^6.1.2",
-    "tslint-sonarts": "^1.9.0",
-    "typescript": "^3.9.5",
-    "webpack": "^4.43.0",
-    "webpack-assets-manifest": "^3.1.1",
-    "webpack-cli": "^3.3.11"
+    "stylelint-config-recommended": "^5.0.0",
+    "stylelint-config-standard": "^22.0.0",
+    "stylelint-scss": "^3.19.0",
+    "svgo": "^2.3.0",
+    "tiny-glob": "^0.2.8",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.2.4"
   },
   "engines": {
-    "node": ">= 10"
+    "node": ">= 14"
   }
 }
