@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -20,17 +20,11 @@
  * IN THE SOFTWARE.
  */
 
-import { identity } from "ramda"
-import { Observable, Subject, asyncScheduler } from "rxjs"
-import {
-  map,
-  observeOn,
-  shareReplay,
-  withLatestFrom
-} from "rxjs/operators"
+import { ObservableInput, Subject, from } from "rxjs"
+import { map, share } from "rxjs/operators"
 
-import { WorkerHandler, watchWorker } from "browser"
-import { translate } from "utilities"
+import { configuration, translation } from "~/_"
+import { WorkerHandler, watchWorker } from "~/browser"
 
 import { SearchIndex, SearchIndexPipeline } from "../../_"
 import {
@@ -41,16 +35,13 @@ import {
 } from "../message"
 
 /* ----------------------------------------------------------------------------
- * Helper types
+ * Types
  * ------------------------------------------------------------------------- */
 
 /**
- * Setup options
+ * Search worker
  */
-interface SetupOptions {
-  index$: Observable<SearchIndex>      /* Search index observable */
-  base$: Observable<string>            /* Location base observable */
-}
+export type SearchWorker = WorkerHandler<SearchMessage>
 
 /* ----------------------------------------------------------------------------
  * Helper functions
@@ -61,7 +52,7 @@ interface SetupOptions {
  *
  * @param data - Search index
  *
- * @return Search index
+ * @returns Search index
  */
 function setupSearchIndex(
   { config, docs, index }: SearchIndex
@@ -69,16 +60,18 @@ function setupSearchIndex(
 
   /* Override default language with value from translation */
   if (config.lang.length === 1 && config.lang[0] === "en")
-    config.lang = [translate("search.config.lang")]
+    config.lang = [
+      translation("search.config.lang")
+    ]
 
   /* Override default separator with value from translation */
-  if (config.separator === "[\s\-]+")
-    config.separator = translate("search.config.separator")
+  if (config.separator === "[\\s\\-]+")
+    config.separator = translation("search.config.separator")
 
   /* Set pipeline from translation */
-  const pipeline = translate("search.config.pipeline")
+  const pipeline = translation("search.config.pipeline")
     .split(/\s*,\s*/)
-    .filter(identity) as SearchIndexPipeline
+    .filter(Boolean) as SearchIndexPipeline
 
   /* Return search index after defaulting */
   return { config, docs, index, pipeline }
@@ -89,51 +82,48 @@ function setupSearchIndex(
  * ------------------------------------------------------------------------- */
 
 /**
- * Set up search web worker
+ * Set up search worker
  *
- * This function will create a web worker to set up and query the search index
- * which is done using `lunr`. The index must be passed as an observable to
+ * This function creates a web worker to set up and query the search index,
+ * which is done using Lunr.js. The index must be passed as an observable to
  * enable hacks like _localsearch_ via search index embedding as JSON.
  *
  * @param url - Worker URL
- * @param options - Options
+ * @param index - Search index observable input
  *
- * @return Worker handler
+ * @returns Search worker
  */
 export function setupSearchWorker(
-  url: string, { index$, base$ }: SetupOptions
-): WorkerHandler<SearchMessage> {
+  url: string, index: ObservableInput<SearchIndex>
+): SearchWorker {
+  const config = configuration()
   const worker = new Worker(url)
 
   /* Create communication channels and resolve relative links */
   const tx$ = new Subject<SearchMessage>()
   const rx$ = watchWorker(worker, { tx$ })
     .pipe(
-      withLatestFrom(base$),
-      map(([message, base]) => {
+      map(message => {
         if (isSearchResultMessage(message)) {
-          for (const { article, sections } of message.data) {
-            article.location = `${base}/${article.location}`
-            for (const section of sections)
-              section.location = `${base}/${section.location}`
-          }
+          for (const result of message.data)
+            for (const document of result)
+              document.location = `${config.base}/${document.location}`
         }
         return message
       }),
-      shareReplay(1)
+      share()
     )
 
   /* Set up search index */
-  index$
+  from(index)
     .pipe(
-      map<SearchIndex, SearchSetupMessage>(index => ({
+      map<SearchIndex, SearchSetupMessage>(data => ({
         type: SearchMessageType.SETUP,
-        data: setupSearchIndex(index)
-      })),
-      observeOn(asyncScheduler)
+        data: setupSearchIndex(data)
+      }))
     )
       .subscribe(tx$.next.bind(tx$))
 
-  /* Return worker handler */
+  /* Return search worker */
   return { tx$, rx$ }
 }
