@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -20,49 +20,49 @@
  * IN THE SOFTWARE.
  */
 
-import { Observable, OperatorFunction, combineLatest, pipe } from "rxjs"
 import {
+  Observable,
+  Subject,
+  animationFrameScheduler,
+  combineLatest,
+  defer,
+  of
+} from "rxjs"
+import {
+  bufferCount,
+  combineLatestWith,
   distinctUntilChanged,
+  distinctUntilKeyChanged,
   filter,
   map,
+  observeOn,
   shareReplay,
   startWith,
-  switchMap,
-  withLatestFrom
+  switchMap
 } from "rxjs/operators"
 
+import { feature } from "~/_"
+import { resetHeaderState, setHeaderState } from "~/actions"
 import {
   Viewport,
-  getElement,
-  watchViewportAt
-} from "browser"
+  watchElementSize,
+  watchToggle
+} from "~/browser"
 
-import { useComponent } from "../../_"
-import {
-  applyHeaderType,
-  watchHeader
-} from "../react"
+import { Component } from "../../_"
+import { Main } from "../../main"
 
 /* ----------------------------------------------------------------------------
  * Types
  * ------------------------------------------------------------------------- */
 
 /**
- * Header type
- */
-export type HeaderType =
-  | "site"                             /* Header shows site title */
-  | "page"                             /* Header shows page title */
-
-/* ------------------------------------------------------------------------- */
-
-/**
  * Header
  */
 export interface Header {
-  type: HeaderType                     /* Header type */
-  sticky: boolean                      /* Header stickyness */
   height: number                       /* Header visible height */
+  sticky: boolean                      /* Header stickyness */
+  hidden: boolean                      /* User scrolled past threshold */
 }
 
 /* ----------------------------------------------------------------------------
@@ -70,11 +70,65 @@ export interface Header {
  * ------------------------------------------------------------------------- */
 
 /**
+ * Watch options
+ */
+interface WatchOptions {
+  viewport$: Observable<Viewport>      /* Viewport observable */
+}
+
+/**
  * Mount options
  */
 interface MountOptions {
-  document$: Observable<Document>      /* Document observable */
   viewport$: Observable<Viewport>      /* Viewport observable */
+  header$: Observable<Header>          /* Header observable */
+  main$: Observable<Main>              /* Main area observable */
+}
+
+/* ----------------------------------------------------------------------------
+ * Helper functions
+ * ------------------------------------------------------------------------- */
+
+/**
+ * Compute whether the header is hidden
+ *
+ * If the user scrolls past a certain threshold, the header can be hidden when
+ * scrolling down, and shown when scrolling up.
+ *
+ * @param options - Options
+ *
+ * @returns Toggle observable
+ */
+function isHidden({ viewport$ }: WatchOptions): Observable<boolean> {
+  if (!feature("header.autohide"))
+    return of(false)
+
+  /* Compute direction and turning point */
+  const direction$ = viewport$
+    .pipe(
+      map(({ offset: { y } }) => y),
+      bufferCount(2, 1),
+      map(([a, b]) => [a < b, b] as const),
+      distinctUntilKeyChanged(0)
+    )
+
+  /* Compute whether header should be hidden */
+  const hidden$ = combineLatest([viewport$, direction$])
+    .pipe(
+      filter(([{ offset }, [, y]]) => Math.abs(y - offset.y) > 100),
+      map(([, [direction]]) => direction),
+      distinctUntilChanged()
+    )
+
+  /* Compute threshold for hiding */
+  const search$ = watchToggle("search")
+  return combineLatest([viewport$, search$])
+    .pipe(
+      map(([{ offset }, search]) => offset.y > 400 && !search),
+      distinctUntilChanged(),
+      switchMap(active => active ? hidden$ : of(false)),
+      startWith(false)
+    )
 }
 
 /* ----------------------------------------------------------------------------
@@ -82,43 +136,71 @@ interface MountOptions {
  * ------------------------------------------------------------------------- */
 
 /**
- * Mount header from source observable
+ * Watch header
  *
+ * @param el - Header element
  * @param options - Options
  *
- * @return Operator function
+ * @returns Header observable
+ */
+export function watchHeader(
+  el: HTMLElement, options: WatchOptions
+): Observable<Header> {
+  return defer(() => {
+    const styles = getComputedStyle(el)
+    return of(
+      styles.position === "sticky" ||
+      styles.position === "-webkit-sticky"
+    )
+  })
+    .pipe(
+      combineLatestWith(watchElementSize(el), isHidden(options)),
+      map(([sticky, { height }, hidden]) => ({
+        height: sticky ? height : 0,
+        sticky,
+        hidden
+      })),
+      distinctUntilChanged((a, b) => (
+        a.sticky === b.sticky &&
+        a.height === b.height &&
+        a.hidden === b.hidden
+      )),
+      shareReplay(1)
+    )
+}
+
+/**
+ * Mount header
+ *
+ * This function manages the different states of the header, i.e. whether it's
+ * hidden or rendered with a shadow. This depends heavily on the main area.
+ *
+ * @param el - Header element
+ * @param options - Options
+ *
+ * @returns Header component observable
  */
 export function mountHeader(
-  { document$, viewport$ }: MountOptions
-): OperatorFunction<HTMLElement, Header> {
-  return pipe(
-    switchMap(el => {
-      const header$ = watchHeader(el, { document$ })
+  el: HTMLElement, { header$, main$ }: MountOptions
+): Observable<Component<Header>> {
+  const internal$ = new Subject<Main>()
+  internal$
+    .pipe(
+      distinctUntilKeyChanged("active"),
+      combineLatestWith(header$),
+      observeOn(animationFrameScheduler)
+    )
+      .subscribe(([{ active }, { hidden }]) => {
+        if (active)
+          setHeaderState(el, hidden ? "hidden" : "shadow")
+        else
+          resetHeaderState(el)
+      })
 
-      /* Compute whether the header should switch to page header */
-      const type$ = useComponent("main")
-        .pipe(
-          map(main => getElement("h1, h2, h3, h4, h5, h6", main)!),
-          filter(hx => typeof hx !== "undefined"),
-          withLatestFrom(useComponent("header-title")),
-          switchMap(([hx, title]) => watchViewportAt(hx, { header$, viewport$ })
-            .pipe(
-              map(({ offset: { y } }) => {
-                return y >= hx.offsetHeight ? "page" : "site"
-              }),
-              distinctUntilChanged(),
-              applyHeaderType(title)
-            )
-          ),
-          startWith<HeaderType>("site")
-        )
-
-      /* Combine into single observable */
-      return combineLatest([header$, type$])
-        .pipe(
-          map(([header, type]): Header => ({ type, ...header })),
-          shareReplay(1)
-        )
-    })
-  )
+  /* Connect to long-living subject and return component */
+  main$.subscribe(main => internal$.next(main))
+  return header$
+    .pipe(
+      map(state => ({ ref: el, ...state }))
+    )
 }
