@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -21,7 +21,6 @@
  */
 
 export * from "./clipboard"
-export * from "./dialog"
 export * from "./instant"
-export * from "./keyboard"
 export * from "./search"
+export * from "./version"
