@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -20,19 +20,7 @@
  * IN THE SOFTWARE.
  */
 
-import { h } from "utilities"
-
-/* ----------------------------------------------------------------------------
- * Data
- * ------------------------------------------------------------------------- */
-
-/**
- * CSS classes
- */
-const css = {
-  wrapper: "md-typeset__scrollwrap",
-  table:   "md-typeset__table"
-}
+import { h } from "~/utilities"
 
 /* ----------------------------------------------------------------------------
  * Functions
@@ -43,14 +31,12 @@ const css = {
  *
  * @param table - Table element
  *
- * @return Element
+ * @returns Element
  */
-export function renderTable(
-  table: HTMLTableElement
-) {
+export function renderTable(table: HTMLElement): HTMLElement {
   return (
-    <div class={css.wrapper}>
-      <div class={css.table}>
+    <div class="md-typeset__scrollwrap">
+      <div class="md-typeset__table">
         {table}
       </div>
     </div>
